+++ app/src/test/java/org/simple/clinic/home/overdue/search/OverdueSearchUpdateTest.kt
@@ -1,5 +1,6 @@
 package org.simple.clinic.home.overdue.search
 
+import androidx.paging.PagingData
 import com.spotify.mobius.test.NextMatchers.hasEffects
 import com.spotify.mobius.test.NextMatchers.hasModel
 import com.spotify.mobius.test.NextMatchers.hasNoEffects
@@ -8,7 +9,9 @@ import com.spotify.mobius.test.UpdateSpec
 import com.spotify.mobius.test.UpdateSpec.assertThatNext
 import org.junit.Test
 import org.simple.clinic.home.overdue.search.OverdueSearchQueryValidator.Result.Valid
+import org.simple.sharedTestCode.TestData
 import java.time.LocalDate
+import java.util.UUID
 
 class OverdueSearchUpdateTest {
 
@@ -54,5 +57,31 @@ class OverdueSearchUpdateTest {
             hasNoModel(),
             hasEffects(AddQueryToOverdueSearchHistory(searchQuery), SearchOverduePatients(searchQuery, date))
         ))
+  }
+
+  @Test
+  fun `when overdue search results are loaded, then update the model`() {
+    val facilityUuid = UUID.fromString("7dba16a0-1090-41f6-8e0c-0d97989de898")
+    val overdueAppointments = listOf(TestData.overdueAppointment(
+        facilityUuid = facilityUuid,
+        name = "Anish Acharya",
+        patientUuid = UUID.fromString("37259e96-e757-4608-aeae-f1a20b088f09")
+    ), TestData.overdueAppointment(
+        facilityUuid = facilityUuid,
+        name = "Anirban Dar",
+        patientUuid = UUID.fromString("53659148-a157-4aa4-92fb-c0a7991ae872")
+    ))
+
+    val overdueSearchResult = PagingData.from(overdueAppointments)
+
+    updateSpec
+        .given(defaultModel)
+        .whenEvent(OverdueSearchResultsLoaded(overdueSearchResult))
+        .then(
+            assertThatNext(
+                hasModel(defaultModel.overdueSearchResultsLoaded(overdueSearchResult)),
+                hasNoEffects()
+            )
+        )
   }
 }